Move CNFOnTheFly into HOL folder
