+++ Inferences/HOL/CNFOnTheFly.cpp
